Create README.md
This project is done when I am serving military service in South Korea. I tried to be familiar with python tkinter GUI module. I made simple Hospital client and server system. Client side is for doctor diagnosis patient and server side is for Hospital office to store records and manipulate patient list. 

Client side can create a patient and add to the list and write prescription.
Server side receives a patient information and prescription. Then, add to the list and check patient is first time visit or not. Next, store all the records of previous patient's data.